+++ core/src/main/java/greencity/webcontroller/ManagementUserController.java
@@ -1,4 +1,4 @@
-package greencity.controller;
+package greencity.webcontroller;
 
 import greencity.entity.User;
 import greencity.service.UserService;
@@ -27,7 +27,6 @@ public class ManagementUserController {
      * @return View template path {@link String}.
      * @author Vasyl Zhovnir
      */
-
     @GetMapping("")
     public String getAllUsers(Model model,
                               @RequestParam(defaultValue = "0") int page,
